+++ lvx frontend/src/pages/about/about.css
